Added some more stuff into the standard package.
NaiveParticlePool contains even more dead code, mostly because a rework
of the linking system needs to be done before concurrent access can be
done for the Living iterator.

The standard system is doing well, and comming up will be a few
implementations! (fire being at the front)

For that to happen, though, the openGL part must be ready for action.